services/graph-engine/requirements.txt
# API
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Base de datos
supabase==2.4.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Validación
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Grafo
networkx==3.2.1

# Data
numpy==1.26.4
pandas==2.2.1

# Tests
pytest==8.1.1
httpx==0.25.2

# Logging
loguru==0.7.2